// src/types/roleType.ts
// Interface untuk area
export type RoleType = {
    id: string;
    name: string; // Nama area
    permissions: { id: number; name: string }[]; // Array of objects untuk lokasi
    createdAt: string; // Tanggal pembuatan (format ISO string)
    updatedAt: string; // Tanggal pembaruan terakhir (format ISO string)
  };
  